ADDED GRADIENT CLIPPING AND WEIGHT DECAY
Now the recurrent models have weight decay an the gradient is
clipped during the backward pass (not the final gradient).

Also the training and testing state of the models is now setted
properly.

There is a problem with the dataset loader, the samples are not
random as expected... This results in poor training and must be
solved quickly...